Update test page to use redirections within the same tab.
This also stops sending back results via window.opener and instead sends
them via localStorage. To achieve this, I propagate the results through all
of the test pages via a query parameter. We will want to move this to
server-side storage since the number of test results will grow quite
long and lead to URLs in the 10s of thousands of characters.