Update Request Blocklist test page to provide downloadable results (#369)
It turns out that using the Tracker Blocking test page to test the Request
Blocklist feature isn't practical for the integration tests on all
platforms. Let's update the Request Blocklist test page therefore, to provide a
downloadable results file (and `window.results` variable), so that test page can
be used for the integration tests instead.